Fix the "Cannot read property 'action' of undefined" errors [0.3.1]
Problem: the foreground (swap.js) called the background (toggle.js) with
{action: 'init'} with the idea to probe whether or not if it should swap
the images in the current tab. If the tab was not active the background
would have sent an empty response which was causing the error.

Solution: Treat the messaging between the background (toggle.js) and the
foreground (swap.js) more carefully:
 - always check if there is a response/request
 - send {action: 'none'} when the foreground should do nothing
 - do nothing by default and only do work if the signal is known